google-drive-invoice-analyzer/requirements.txt
mcp[cli]>=1.5.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
PyPDF2>=3.0.0
watchdog>=3.0.0
pandas>=2.0.0
pydantic>=2.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-dotenv>=1.0.0
